pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kappa"
version = "0.1.0"
description = "Autonomous self-healing agent harness with deterministic guardrails"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "langgraph>=0.2.0",
    "docker>=7.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kappa"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
